Replace AsyncTask with AsyncTaskLoader
This will handle the case that our app is recreated and we want to use the same background threat rather than create a new one. Also a big issue in this exercise was that I accidently used the entered search text to put into the bundle. BIG ISSUE because that isn't the FINAL constructed URL. This is an issue here because the bundle is being used by the loader to get the necessary url to query. Note: The communication between the main thread's menu button "search" to the loader itself is by transferring data through the bundle. Look at function definition of LoaderManager.initLoader which requries a Bundle.