cache data
Put the data in a string variable and we deliver it directly to loader rather than forcing a new load. Note: we override deliverResult to make sure we attach the data to mGithubJson or else after the first time force a load that it won't know that we already have loaded once.